Moved notifications enums to package notification
