+++ grails-app/domain/com/miniasaaslw/domain/notification/Notification.groovy
@@ -2,7 +2,8 @@ package com.miniasaaslw.domain.notification
 
 import com.miniasaaslw.domain.customer.Customer
 import com.miniasaaslw.entity.BaseEntity
-import com.miniasaaslw.entity.enums.NotificationPriority
+import com.miniasaaslw.entity.enums.notification.NotificationPriority
+import com.miniasaaslw.entity.enums.notification.NotificationType
 
 class Notification extends BaseEntity {
 
@@ -13,6 +14,8 @@ class Notification extends BaseEntity {
     String message
 
     String url
+
+    NotificationType type
 
     NotificationPriority priority
 }
